Alireza 26 02 2020 authentication

+++ src/tasks/tasks.service.ts
@@ -7,6 +7,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
  
 @Injectable()
 export class TasksService {
@@ -16,12 +17,13 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) { }
 
-    async getTasks(filterDto : GetFilterTaskDto) : Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto : GetFilterTaskDto , user : User) : Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto , user);
     }
 
-    async getTaskById(id : number) :Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id : number , user : User) :Promise<Task> {
+
+        const found = await this.taskRepository.findOne({ where : { id , userId: user.id } });
 
         if(!found){
             throw new NotFoundException(`Task with id "${id}" not found`);
@@ -30,22 +32,22 @@ export class TasksService {
         return found;
     }
 
-    async createTask(createTaskDto : CreateTaskDto) : Promise<Task>{
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto : CreateTaskDto , user : User) : Promise<Task>{
+        return this.taskRepository.createTask(createTaskDto , user);
     }
 
-    async deleteTask(id : number) :Promise<void> {
+    async deleteTask(id : number , user : User) :Promise<void> {
 
-        const result = this.taskRepository.delete(id);
+        const result = this.taskRepository.delete({ id , userId : user.id });
 
         if((await result).affected == 0){
             throw new NotFoundException(`Task with ID "${id}" not found`);
         }
     }
 
-    async updateTaskSatus(id : number , status : TaskStatus) : Promise<Task>{
+    async updateTaskSatus(id : number , status : TaskStatus , user : User) : Promise<Task>{
 
-        const task = this.getTaskById(id);
+        const task = this.getTaskById(id , user);
         (await task).status = status;
         (await task).save();
